If -Force is specified, use the newest usable Python interpreter.

+++ PythonProvider/PythonProvider.cs
@@ -135,12 +135,26 @@ namespace PythonProvider
             }
             else if (usableinstalls.Count > 1)
             {
-                request.Warning("Multiple installed Python interpreters could satisfy this request:");
-                foreach (var install in usableinstalls)
+                if (request.GetOptionValue("Force") == "True")
                 {
-                    request.Warning("  Python version '{0}' at '{1}'", install.python_version, install.exe_path);
+                    PythonInstall greatest = usableinstalls[0];
+                    foreach (var candidate in usableinstalls)
+                    {
+                        if (candidate.python_version.Compare(greatest.python_version) > 0)
+                            greatest = candidate;
+                    }
+                    package.Install(greatest, request);
                 }
-                request.Error(ErrorCategory.NotSpecified, package.name, "Please select a Python to install to, using e.g. -PythonVersion 3.2 or -PythonLocation c:\\python32\\python.exe");
+                else
+                {
+                    request.Warning("Multiple installed Python interpreters could satisfy this request:");
+                    foreach (var install in usableinstalls)
+                    {
+                        request.Warning("  Python version '{0}' at '{1}'", install.python_version, install.exe_path);
+                    }
+                    request.Warning("Please select a Python to install to, using e.g. -PythonVersion 3.2 or -PythonLocation c:\\python32\\python.exe");
+                    request.Error(ErrorCategory.NotSpecified, package.name, "Not enough information to select a Python interpreter for the install");
+                }
             }
         }
     }
